Don't use popups, they are blocked in browsers
Basically all modern browsers (or at least Firefox and Chrome) block
popups (i.e. windows opened by `window.open`) by default. So this makes
for a suboptimal user experience.

The way around this is to add the final URL and target="_blank" to the
link by default and let the browser execute the default action.

When creating the link, it is necessary to differentiate between the
different actions. We currently could use the `label`-key, however, at
some point we might want to localize that 'label' key. Hence an `actionId`
is added to reliably identify the action.