Limit the width of the displayed links
The name of a page or an media file might be very(!) long and cause the
'view' link to flow to the next line. However, there it might be
unclickable because it might be below the next line's list item.

This solution limits the width of the item and enforces linebreaks early
enough, so that the 'view'-link can still flow to the right of the item
and not below it.