Fix: Use safe explode (sexplode)
Error parsing answer:

Warning: Undefined array key 1 in /home/nethouse/wiki/lib/plugins/multiorphan/action/multiorphan.php on line 437

This is a PHP warning, indicating that the ``explode()`` function, when used with '#' as the delimiter on the variable ``$id``, did not find a '#' character. As a result, the resulting array only has one element (index 0), and when you try to access index 1 (i.e., ``$hash``), it's "undefined.

This ensures the array always has a minimum number of elements, filling missing ones with a default value (null by default).